backtest/crawler/get_api_key.py
import os

from dotenv import load_dotenv
from py_clob_client_v2.client import ClobClient

load_dotenv()

host: str = "https://clob.polymarket.com"
key: str = os.getenv("POLYMARKET_PRIVATE_KEY")  # This is your Private Key.
chain_id: int = 137  # No need to adjust this
POLYMARKET_PROXY_ADDRESS: str = os.getenv(
    "POLYMARKET_FUNDER"
)  # This is the address you deposit/send USDC to to FUND your Polymarket account.

# Select from the following 3 initialization options to matches your login method, and remove any unused lines so only one client is initialized.

### Initialization of a client using a Polymarket Proxy associated with an Email/Magic account. If you login with your email use this example.
client = ClobClient(
    host, key=key, chain_id=chain_id, signature_type=1, funder=POLYMARKET_PROXY_ADDRESS
)

### Initialization of a client using a Polymarket Proxy associated with a Browser Wallet(Metamask, Coinbase Wallet, etc)
client = ClobClient(
    host, key=key, chain_id=chain_id, signature_type=2, funder=POLYMARKET_PROXY_ADDRESS
)

### Initialization of a client that trades directly from an EOA.
client = ClobClient(host, key=key, chain_id=chain_id)

print(client.derive_api_key())
