pyproject.toml
[project]
name = "dr-manhattan"
version = "0.0.2"
description = "CCXT-style unified API for prediction markets with MCP server"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "guzus" }
]
keywords = ["prediction-markets", "polymarket", "opinion", "trading", "api", "ccxt"]
dependencies = [
    "requests>=2.31.0",
    "websockets>=15.0.1",
    "python-socketio[asyncio_client]>=5.11.0",
    "python-dotenv>=1.0.0",
    "eth-account>=0.11.0",
    "opinion-clob-sdk>=0.4.3",
    "pandas>=2.0.0",
    "boto3>=1.42.14",
    "pyarrow>=22.0.0",
    "matplotlib>=3.10.8",
    "rich>=14.2.0",
    "predict-sdk>=0.0.8",
    "cryptography>=42.0.0",
    "scikit-learn>=1.8.0",
    "py-clob-client-v2>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/guzus/dr-manhattan"
Repository = "https://github.com/guzus/dr-manhattan"
Issues = "https://github.com/guzus/dr-manhattan/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = [
    "N818",  # Exception names don't need Error suffix for backward compatibility
    "E501",  # Line too long (handled by black where possible)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/frontend",
    "/backtest",
    "/wiki",
    "/scripts",
    "/examples",
    "*.ipynb",
]

[tool.hatch.build.targets.wheel]
packages = ["dr_manhattan"]

[project.optional-dependencies]
mcp = [
    "mcp>=0.9.0",
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
dr-manhattan-mcp = "dr_manhattan.mcp.server:run"
dr-manhattan-mcp-sse = "dr_manhattan.mcp.server_sse:run_sse"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=24.0.0",
    "ruff==0.14.2",
    "twine>=6.0.0",
    "pre-commit>=4.5.1",
]
